app/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["mattyocode"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "3.2.7"
djangorestframework = "3.12.4"
psycopg2-binary = "2.9.1"
gunicorn = "20.1.0"
whitenoise = "5.2.0"
django-cors-headers = "^3.10.0"
djangorestframework-simplejwt = "^5.0.0"
requests = "^2.26.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
pytest-django = "4.4.0"
pytest-factoryboy = "^2.1.0"
pytest-cov = "^2.12.1"
flake8 = "^3.9.2"
black = "21.6b0"
isort = "5.9.1"
pytest-mock = "^3.6.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
multi_line_output=3
include_trailing_comma=true